backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
alpaca-trade-api
backtrader 
matplotlib
sqlalchemy
psycopg2-binary